fix notification ID counter type mismatch in settings storage
Changed setLong to setInt for NOTIFICATION_ID_COUNTER (2 locations) to match
corresponding getInt() calls. Prevents notification ID collisions and incorrect
counter resets.

+++ app/src/main/java/de/kaffeemitkoffein/tinyweatherforecastgermany/WeatherSettings.java
@@ -987,7 +987,7 @@ public class WeatherSettings {
         pref_editor.putInt(PREF_NOTIFICATION_IDENTIFIER,b);
         pref_editor.apply();
          */
-        DataStorage.setLong(context,DataStorage.DATASTORAGE_NOTIFICATION_IDENTIFIER,b);
+        DataStorage.setInt(context,DataStorage.DATASTORAGE_NOTIFICATION_IDENTIFIER,b);
         return i;
     }
 
@@ -1000,7 +1000,7 @@ public class WeatherSettings {
             pref_editor.putInt(PREF_DISPLAY_OVERVIEWCHART,NotificationIDRange.MIN);
             pref_editor.apply();
              */
-            DataStorage.setLong(context,DataStorage.DATASTORAGE_NOTIFICATION_IDENTIFIER,NotificationIDRange.MIN);
+            DataStorage.setInt(context,DataStorage.DATASTORAGE_NOTIFICATION_IDENTIFIER,NotificationIDRange.MIN);
         }
     }
